% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_gapminder.R
\docType{data}
\name{my_gapminder}
\alias{my_gapminder}
\title{Data from Gapminder}
\format{a data frame with 1704 observations and 6 variables:
\describe{
\item{country}{country}
\item{continent}{continent}
\item{year}{year}
\item{lifeExp}{total population}
\item{gdpPercap}{per-capita GDP}
}}
\source{
\url{https://www.gapminder.org/data/}
}
\usage{
my_gapminder
}
\description{
An excerpt of the data available at Gapminder.org. For each of 142 countries,
the package provides values for life expectancy, GDP per capita, and population,
every five years, from 1952 to 2007.
}
\keyword{datasets}
