% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{My linear model function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{a \code{formula} class object, a symbolic description of the
model to be fitted.}

\item{data}{input data frame.}
}
\value{
a summary table:
\code{Estimate}, the beta value;
\code{Std. Error}, the standard error;
\code{t value} the t value;
\code{Pr(>|t|)}, the p value.
}
\description{
This function fits a linear model in R.
}
\examples{
data(mtcars)
my_lm(mpg ~ hp + wt, data = mtcars)
data(my_gapminder)
my_lm(pop ~ gdpPercap, data = my_gapminder)

}
\keyword{prediction}
