% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t_test.R
\name{my_t_test}
\alias{my_t_test}
\title{My t-test function}
\usage{
my_t_test(x, alternative, mu)
}
\arguments{
\item{x}{a non-empty numeric vector of data values.}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided", "greater" or "less".}

\item{mu}{a number indicating the true value of the mean
(or difference in means if you are performing a two sample test).}
}
\value{
the t test results of x:
a list includes,
\code{test_stat}, the value of the test statistic;
\code{df}, the degrees of freedom;
\code{alternative} a character string describing the alternative hypothesis;
\code{p_val}, the p-value for the test.
}
\description{
This function performs a one sample t-test in R.
}
\examples{
my_t_test(x = c(1:200), alternative = "less", mu = 25)
my_t_test(c(1:100), greater, 10)

}
\keyword{inference}
